api/requirements.txt
# AI Gateway API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
azure-search-documents>=11.4.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
